+++ changelog.md
@@ -9,11 +9,12 @@
   use in a compilation view.
 * Better handling of ANSI colors in Fennel Proto REPL.
 * `ob-fennel` now can send Emacs Lisp hash tables to Fennel.
-* major update to the fennel-proto-repl - new protocol version,
+* major update to the `fennel-proto-repl` - new protocol version,
   changes in user-input handling, ability to extend with custom
   operations.
 * Add Flymake support via `fennel-ls`
 * Add dynamic font-locking support to the `fennel-proto-repl` mode
+* Add project integration to the `fennel-proto-repl` mode
 
 ## 0.9.1 / 2024-04-24